// src/apps/kroncl/wm/types.ts
export type UnitType = 'product' | 'service';
export type UnitStatus = 'active' | 'inactive';
export type InventoryType = 'tracked' | 'untracked';
export type TrackingDetail = 'batch' | 'serial';
export type TrackedType = 'fifo' | 'lifo';
export type CurrencyType = 'RUB';

export interface CatalogUnit {
    id: string;
    name: string;
    comment: string | null;
    type: UnitType;
    status: UnitStatus;
    inventory_type: InventoryType;
    tracking_detail?: TrackingDetail | null;
    tracked_type: TrackedType | null;
    unit: string;
    sale_price: number;
    purchase_price: number | null;
    currency: CurrencyType;
    category_id: string;
    metadata: Record<string, any> | null;
    created_at: string;
    updated_at: string;
}